Update Polder ancestry
I added the Reactive Tumble purchased trait, which is in the official source book but wasn't in `polder.ts`. I kept the traits in order of ascending purchase cost, and renumbered the Id of the more expensive traits. Let me know if I should instead put the new trait at the end and/or give it id `'polder-feature-3-6'` to leave the two more expensive traits with the same id.